user profile + profile update

+++ src/routes/projects.js
@@ -74,13 +74,19 @@ router.get("/:project_id",
 router.post("/:project_id/remove",
     (req, res) => {
         let id = req.params.project_id;
-        Project.findByIdAndRemove(id,
-            function (err, project) {
-                if (err) {
-                    res.sendStatus(404);
-                    return;
-                }
-                res.redirect("/projects");
-            }) //.catch(err => res.sendStatus(500));
+        if (id == req.user.id) {
+            Project.findByIdAndRemove(id,
+                function (err, project) {
+                    if (err) {
+                        //  res.sendStatus(404);
+                        res.render("error_page", {
+                            error: res.statusMessage
+                        });
+                        return;
+                    }
+                    res.redirect("/projects");
+                })
+        }
+        //.catch(err => res.sendStatus(500));
     });
 module.exports = router;
